V4L2: uvcvideo: allow switching frame formats with allocated buffers
Currently the uvcvideo V4L2 driver blocks attempts to change the frame
format, using the VIDIOC_S_FMT ioctl(), when buffers are already allocated.
This was logical until now, but recently support for the
VIDIOC_CREATE_BUFS ioctl() has been added to the driver. With that ioctl()
buffers for several formats can be allocated for a camera instance
simultaneously. This is usually done to be able to switch between formats
by just stopping streaming, setting a new format, queuing new buffers and
restarting streaming. To enable this we remove the restriction and only
block VIDIOC_S_FMT during a running streaming. Note, that this isn't quite
safe, because there is currently no way to verify, that queued buffers are
suitable for the configured format. Such checks should be added in the
future.

Change-Id: I0ddb666e38a11d82dae7042a449bd6389519494c
Signed-off-by: Guennadi Liakhovetski <g.liakhovetski@gmx.de>