i915: Expose PMU for Observation Architecture
Gen graphics hardware can be set up to periodically write snapshots of
performance counters into a circular buffer and this patch exposes that
capability to userspace via the perf interface.

This introduces a new PERF_PMU_CAP_IS_DEVICE flag that allows us to
convey to events/core that our pmu does not relate to the cpu. This in
turn lets us bypass the perf_paraniod_cpu() || CAP_SYS_ADMIN check when
opening an event associated with a specific cpu id but no specific pid.
In our case that doesn't imply we will expose anything about the
different processes that may run on that cpu. This driver builds on the
drm driver's authentication scheme to determine if a process has
permission to collect metrics for a specific gpu context.  Only root can
collect metrics across all gpu contexts.

Only Haswell is supported currently.

Signed-off-by: Robert Bragg <robert@sixbynine.org>