drm/i915: Add runtime eu/slice fuse checks for Broadwell
Userspace will need information about the EU, slice and sub slice
configuration to be able to normalize some OA metrics. Some of this can
be determined based on the device id, but we will also need to expose
further information such as the slice and subslice mask which can't be
derived from the device id. Its convenient for userspace if we can
report this information consistently for Gen 8+

Signed-off-by: Robert Bragg <robert@sixbynine.org>