drm/i915: Expose a subslices enabled mask to userspace
Certain performance counter configurations depend on specific subslices
being enabled. Userspace needs the mask of enabled subslices too (not
just the count) to be able to process the raw performance counter
reports and ignore invalid data.

This adds an I915_PARAM_SUB_SLICE_MASK param where a value of zero
implies the query is not supported.

Signed-off-by: Robert Bragg <robert@sixbynine.org>