drm/i915: Introduce mode for asynchronous capture of OA counters
This patch introduces fields for supporting asynchronous capture of OA counter
snapshot.
There may be usecases wherein we need more than periodic OA capture mode which
is supported by perf_event currently. We may need to insert the commands into
the ring to dump the OA counter snapshots at some asynchronous points during
workload execution.
This mode is primarily used for two usecases:
    - Ability to capture system wide metrics. The reports captured should be
      able to be mapped back to individual contexts.
    - Ability to inject tags for work, into the reports. This provides
      visibility into the multiple stages of work within single context.

The asynchronous reports generated in this way (using commands), will be
forwarded to userspace after appending a footer, which will have this metadata
information. This will enable the usecases mentioned above.

This patch introduces an additional field in the oa attr structure for supporting this
type of capture. The data thus captured needs to be stored in a seperate buffer,
which will be different from the buffer used otherwise for periodic OA capture
mode. Again this buffer address will not need to be mapped to OA unit register
addresses such as OASTATUS1, OASTATUS2 and OABUFFER.

The subsequent patches introduce the data structures and mechanism for
forwarding reports to userspace, and mechanism for inserting corresponding
commands into the ringbuffer.

Change-Id: Ib3759df6b9fc47f5430862d19bd9d9bb5112b3f9
Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>