drm/i915: Add mechanism for forwarding async OA counter snapshots through perf
This patch adds the mechanism for forwarding the asynchronous OA snapshots
through the perf event interface.

The data captured will be organized into nodes. Each node has the field for OA
report alongwith metadata information such as ctx_id, pid, etc. The metadata
information can be extended to provided any additional information.
The data is organized to have a queue header at beginning, which will have
information about size, data offset, number of nodes captured etc.

Each node of data collected is forwarded as a seperate perf sample.
A single snapshot will have two fields. First is the raw report and second
field is a footer with metadata corresponding to snapshot such as ctx_id, pid.
The size of the raw report is the one specified during event init.

In order to track whether the gpu has completed processing the node, a field
pertaining to corresponding gem request structure is added into the node info.
The request structure is expected to be assigned whenever the gpu command is
submitted. By doing this, a reference is taken for the request, so that it
doesn't disappear till the time we've checked for its status while forwarding
the node data.

While forwarding the samples, we check whether the gem request is completed and
dereference the corresponding request. The need to dereference the request
necessitates a worker here, which will be scheduled when the hrtimer triggers.

While flushing the samples, we have to wait for the requests already scheduled,
before forwarding the samples. This wait can be done in a lockless fashion.

Change-Id: I2aa2a163d9c6034dee7fbb4d8fc380a64708b47d
Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>