drm/i915: Handle event stop and destroy for commands in flight
In the periodic OA sampling mode, the event stop would stop forwarding samples
to userspace, and disables OA synchronously. The buffer is destroyed eventually
in event destroy callback. But when we have in flight RPC commands scheduled on
GPU (like in this case), the handling of OA disabling and buffer destruction
has to take those into account.

Now, the event stop would just set the event state, and stop forwarding data to
userspace. From userspace perspective, for all purposes, the event sampling is
stopped. This is true for periodic OA samples also. The OA unit is not disabled
here, since there may be RPC commands scheduled on GPU.

A subsequent event start (which comes without event destroy) may need to wait
for commands in flight and release the references, before starting sampling
again.

The event destroy should now be triggering the destruction of destination
buffer.
One of the approaches to implement this is to take an active reference on the
destination buffer during the event stop. We can then just drop our local
reference during event destroy, and rely on gem reference tracking to destroy
the buffer whenever GPU commands scheduled are finished.
But the problem here is that, here, we are storing the gem request pointers,
which would be referenced while submitting commands, in the destination buffer
itself. If the destination buffer disappears via gem reference tracking, we
won't have chance to dereference those requests.
The ideal solution here would be to have a callback when the last request is
finished on GPU, so that we can derefernce the requests (and do some more
cleanup such as releasing forcewake etc). This callback can be set up for
the LRI request, that we can schedule for disabling OA during event stop (or if
LRI is not desirable, some dummy request can be scheduled during event stop).
This functionality is WIP through retire-notification mechanism (by Chris).

For now, the event destroy will not drop local reference to destination buffer.
Instead, it schedules a worker, which will locklessly wait for the requests'
completion, thereafter disabling OA, and destroying the destination buffer.

A subsequent event init would have to wait for previously submitted RPC commands
to complete or return -EBUSY. Currently, for the sake of simplicity, we are
returning -EBUSY if such a case is detected.

The possibility of wait in event_init/event_start is an unavoidable consequence
of asynchronous nature of OA reports by MI_RPC commands.

Change-Id: I850d65936f6b30058f8e474ef3ece4ec295018d9
Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>