drm/i915: Add a new i915 perf event type for collection of timestamps
The current i915 perf _OA_EVENT is specific for collection of OA
counter reports. Since this enables us (and limits us) to the render ring,
we have no means for collection of profiling data pertaining to other rings

To overcome this limitation, we need to have a way and means of capturing
performance data for other rings also (in a non-OA specific way).
This data may encompass timestamps, mmio register values, etc.

For this, a new event type _GEN_EVENT is introduced. This will serve
these purposes, without constraining itself to type of data being dumped
(which may restrict the user to specific ring like in case of OA counters).

In this patch, this event type is used to collect the timestamps around
workloads, on all the rings.
The timestamps will be captured in a gem buffer object. The metadata
information such as ctx_id, pid etc. is maintained in a list, whose each node
has offsets into the gem buffer object for each snapshot captured.
In order to track whether the gpu has completed processing the node,
a field pertaining to corresponding gem request is added, which is tracked for
completion before forwarding timestamps.

Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>