drm/i915: Add support to add execbuffer tags to OA counter reports
This patch enables userspace to specify tags (per workload), provided via
execbuffer ioctl, which could be added to OA reports, to help associate
reports with the corresponding workloads.

There may be multiple stages within a single context, from a userspace
perspective. An ability is needed to individually associate the OA reports
with their corresponding workloads(execbuffers), which may not be possible
solely with ctx_id or pid information. This patch enables such a mechanism.

In this patch, rsvd2 field of execbuffer arguments is being utilized for
passing the tag. A new bitfield in execbuffer flags is introduced in order
to inform kernel of the tag being passed in execbuffer arguments.

v2: Changes (as suggested by Chris):
    - Using upper 32 bits of rsvd1 field for passing tag (as it is a
      natural complement to ctx_id which is there in lower 32 bits of rsvd1
    - Removed the flags field to inform kernel of tag being passed.
    - Better ABI definition.

Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>