drm/i915: Insert commands for capture of OA counters in the ring
This patch adds the routines which insert commands for capturing OA snapshots,
into the ringbuffer of RCS engine.

The command MI_REPORT_PERF_COUNT can be used to capture snapshots of OA
counters. Here, this command is being inserted in the ring at BB boundaries.

Also, we keep a reference of the request in the node header while submitting
the commands. This will be released when we are forwarding the samples to
userspace (or when the event is being stopped/destroyed)

Change-Id: I2d279a2ade413ba892385bdd277d8126411ddac0
Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>