drm/i915: start tracking device_info subslice_mask
Not assuming the subslice mask is uniform across all slices this mask
covers all slices with N bits per slice, where N is the maximum number
of slices for a particular Gen.

Notably this matches the definition of the $SubsliceMask variable that
is currently found in out OA metrics XML files.

Notably this updates the ss_max constant (max number of sub slices) from
4 to 3 because although the FUSE2 register provides a 4 bit mask, gen 9
configurations really only go up to a maximum of 3 slices, with bit 3 in
the FUSE2 subslice disabled mask always set.

Cc: Zhenyu Wang <zhenyuw@linux.intel.com>
Signed-off-by: Robert Bragg <robert@sixbynine.org>