drm/i915: Forward periodic and CS based OA reports sorted acc to timestamps
The periodic reports and the RCS based reports are collected in two
separate buffers. While forwarding to userspace, these have to be sent to
single perf event ringbuffer. From a userspace perspective, it is good to
have the reports in the single buffer in order to their timestamps.
This patch addresses this problem by forwarding the periodic OA reports
with a lower timestamp, whenever we are forwarding the Command streamer
based report.

Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>