drm/i915: Add support for having pid output with OA report
This patch introduces flags and adds support for having pid output with the
OA reports generated through the RCS commands.

When the userspace expresses its interest in listening to the pid through
an oa_attr field during event init, the OA reports generated would have an
additional field appended with the pid information. The patches enables
this framework, which can be expanded upon to introduce further fields in
the OA report metadata.

Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>