squashme: drm/i915: Detect the overflow condition for command stream buf
Add a compile time option for detecting the overflow condition of command
stream buffer, and not overwriting the old entries in such a case.
Also, set a status flag to forward the overflow condition to userspace if
overflow is detected.

Patch to be squashed with base patch (which adds support for capturing command
stream buf based OA reports). Or should it be a seperate patch?

Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>