squashme: drm/i915: Populate ctx ID for periodic OA reports
This adds support for populating the ctx id for the periodic OA reports
when requested through the corresponding property.

For Gen8, the OA reports itself have the ctx ID and it is the one programmed
into HW while submitting workloads. Thus it's retrieved from reports itself.
For Gen7, the OA reports don't have any such field, and we can populate this
field with the last seen ctx ID while sending CS reports.

Patch to be squashed with base patch (which adds support for capturing command
stream buf based OA reports). Or should it be a seperate patch?

Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>